+++ crates/nova-ide/src/framework_db_adapter.rs
@@ -336,7 +336,7 @@ impl ClassIndex {
             path.hash(&mut hasher);
             let text = db.file_content(*file_id);
             text.len().hash(&mut hasher);
-            (text.as_ptr() as usize).hash(&mut hasher);
+            text.as_ptr().hash(&mut hasher);
         }
         let fingerprint = hasher.finish();
